requirements.txt
flask
pymongo
python-dotenv
groq
torch
gunicorn